requirements.txt
streamlit
google-generativeai
pandas
matplotlib
seaborn
scikit-learn
kaggle

ultralytics>=8.0.0
torch>=1.10.0
torchvision>=0.11.0
opencv-python>=4.5.0
pillow
numpy